refactor(arch): route all get_proxmox_adapter calls through services facade and add preflight logging
