+++ src/osx_proxmox_next/defaults.py
@@ -181,7 +181,7 @@ DEFAULT_ISO_DIR = "/var/lib/vz/template/iso"
 
 def detect_iso_storage() -> list[str]:
     """Return ISO directory paths from Proxmox storage pools that support ISO content."""
-    from .services.proxmox_service import get_proxmox_adapter
+    from .services import get_proxmox_adapter
     pve = get_proxmox_adapter()
     dirs: list[str] = []
     res = pve.pvesm("status", "-content", "iso")
