+++ src/osx_proxmox_next/downloader.py
@@ -128,7 +128,7 @@ def download_recovery(
 
 
 def _build_recovery_image(dmg_path: Path, _chunklist_path: Path, dest: Path) -> None:
-    from .services.proxmox_service import get_proxmox_adapter
+    from .services import get_proxmox_adapter
     adapter = get_proxmox_adapter()
     result = adapter.run(["dmg2img", str(dmg_path), str(dest)])
     if not result.ok:
